+++ src/lib/lib.nix
@@ -1,37 +1,24 @@
 { lib, ... }:
 
-# TODO: with mkOptionType
-# NOTE: anything results in infinite recursion?
-
 let
-  nest = lib.fix (
-    nest: times:
-    if times == 0 then
-      lib.types.oneOf [
-        lib.types.bool
-        lib.types.number
-        lib.types.str
-        lib.types.optionType
-        (lib.types.functionTo lib.types.raw)
-      ]
-    else
-      lib.types.oneOf [
-        lib.types.bool
-        lib.types.number
-        lib.types.str
-        lib.types.optionType
-        (lib.types.functionTo lib.types.raw)
-        (lib.types.listOf (nest (times - 1)))
-        (lib.types.attrsOf (nest (times - 1)))
-      ]
-  );
+  recursiveAttrsOf =
+    elemType:
+    lib.types.mkOptionType {
+      name = "recursiveAttrsOf";
+      description = "nested attribute set of ${elemType.description or "values"}";
+      descriptionClass = "noun";
+      check = value: lib.isAttrs value;
+      merge = loc: defs: lib.foldl' lib.recursiveUpdate { } (map (def: def.value) defs);
+    };
 in
 {
   options.flake.lib = lib.mkOption {
-    type = lib.types.attrsOf (nest 8);
+    type = recursiveAttrsOf lib.types.raw;
     default = { };
-    description = lib.literalMD ''
-      `lib` flake output.
-    '';
+    description = "Attribute set of all library functions in the flake";
+  };
+
+  config.flake.lib.types = {
+    inherit recursiveAttrsOf;
   };
 }