+++ CHANGELOG.md
@@ -7,6 +7,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and adheres to [Semantic Versioning](https://semver.org/).
 
+## [1.1.1] - 2026-01-07
+
+### Added
+
+- inputs to special args
+- option descriptions
+- generated options documentation
+
+### Changed
+
+- better handling of `flakeModules` module argument
+
 ## [1.1.0] - 2026-01-05
 
 ### Added
@@ -53,6 +65,7 @@ and adheres to [Semantic Versioning](https://semver.org/).
 
 - everything
 
+[1.1.1]: https://github.com/haras-unicorn/perch/compare/1.1.0...1.1.1
 [1.1.0]: https://github.com/haras-unicorn/perch/compare/1.0.4...1.1.0
 [1.0.4]: https://github.com/haras-unicorn/perch/compare/1.0.3...1.0.4
 [1.0.3]: https://github.com/haras-unicorn/perch/compare/1.0.2...1.0.3